cWorldGenerator speedup - doesn't call GetChunk() anymore, not queueing the chunk it's generating to be loaded recursively. cChunk fix - setting a block to the same value doesn't mark chunk dirty (resulted in un-unloadable chunks)
git-svn-id: http://mc-server.googlecode.com/svn/trunk@279 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/cMonsterConfig.cpp
@@ -105,7 +105,7 @@ void cMonsterConfig::AssignAttributes(cMonster *m, const char* n)
 			m->SetAttackRange  (itr->m_AttackRange);
 			m->SetSightDistance(itr->m_SightDistance);
 			m->SetAttackRate   ((int)itr->m_AttackRate);
-			m->SetMaxHealth    ((int)itr->m_MaxHealth);
+			m->SetMaxHealth    ((short)itr->m_MaxHealth);
 		}
 	}  // for itr - m_pState->AttributesList[]
 }
